better table mapping

+++ oauth/src/main/kotlin/net/wohlfart/charon/repository/AuthUserRepository.kt
@@ -4,7 +4,7 @@ import net.wohlfart.charon.entity.AuthUserDetails
 import org.springframework.data.repository.CrudRepository
 
 
-interface AuthUserRepository: CrudRepository<AuthUserDetails, Long> {
+interface AuthUserRepository: CrudRepository<AuthUserDetails, Int> {
 
     fun findByUsername(username: String): AuthUserDetails
