+++ oauth/src/main/kotlin/net/wohlfart/charon/repository/RegistrationRepository.kt
@@ -4,7 +4,7 @@ import net.wohlfart.charon.entity.UserRegistration
 import org.springframework.data.repository.CrudRepository
 
 
-interface RegistrationRepository: CrudRepository<UserRegistration, Long> {
+interface RegistrationRepository: CrudRepository<UserRegistration, Int> {
 
      fun findByTokenValue(tokenValue: String): UserRegistration
